updated Poi panel according to new design
