+++ react-components/src/architecture/concrete/pointsOfInterest/ads/PointsOfInterestAdsProvider.ts
@@ -54,7 +54,9 @@ export class PointsOfInterestAdsProvider implements PointsOfInterestProvider<Ext
       );
     }
 
-    return result.data.items.map(poiItemToInstance);
+    return await Promise.all(
+      result.data.items.map(async (item) => await poiItemToInstance(item, this._sdk))
+    );
   }
 
   async fetchPointsOfInterest(
@@ -71,7 +73,9 @@ export class PointsOfInterestAdsProvider implements PointsOfInterestProvider<Ext
       );
     }
 
-    return result.data.items.map(poiItemToInstance);
+    return await Promise.all(
+      result.data.items.map(async (item) => await poiItemToInstance(item, this._sdk))
+    );
   }
 
   async deletePointsOfInterest(poiIds: ExternalId[]): Promise<void> {
@@ -152,11 +156,17 @@ export class PointsOfInterestAdsProvider implements PointsOfInterestProvider<Ext
   }
 }
 
-function poiItemToInstance(item: PoiItem): PointsOfInterestInstance<ExternalId> {
+async function poiItemToInstance(
+  item: PoiItem,
+  sdk: CogniteClient
+): Promise<PointsOfInterestInstance<ExternalId>> {
+  const userIdNameMap = await getUserName(item.ownerId, sdk);
   return {
     id: item.externalId,
     properties: {
       title: item.name,
+      ownerId: userIdNameMap,
+      createdTime: item.createdTime,
       positionX: item.position[0],
       positionY: item.position[1],
       positionZ: item.position[2],
@@ -168,4 +178,12 @@ function poiItemToInstance(item: PoiItem): PointsOfInterestInstance<ExternalId> 
       sceneState: item.sceneState
     }
   };
+
+  async function getUserName(ownerId: string, sdk: CogniteClient): Promise<string | undefined> {
+    const profiles = await sdk.profiles.retrieve([{ userIdentifier: ownerId }]);
+
+    const name = profiles.filter(isDefined).map((profile) => profile.displayName ?? undefined);
+
+    return name[0];
+  }
 }
