+++ react-components/src/common/i18n/en/reveal-react-components.json
@@ -29,6 +29,8 @@
   "CREATE": "Create",
   "CROP_BOX": "Crop box",
   "CYLINDER": "Cylinder",
+  "DESCRIPTION": "Description",
+  "DESCRIPTION_PLACEHOLDER": "This is the description",
   "DELETE": "Delete",
   "FIT_VIEW_TOOLTIP": "Fit view",
   "HELP_TOOLTIP": "Help",
@@ -126,6 +128,8 @@
   "TOUCH_NAVIGATION_TITLE": "Touch",
   "TOUCH_SELECT": "Tap to select",
   "TOUCH_ZOOM": "Zoom",
+  "TITLE": "Title",
+  "TITLE_PLACEHOLDER": "This is the title",
   "UNDO": "Undo",
   "WIDGET_WINDOW_CLOSE": "Close",
   "WIDGET_WINDOW_EXPAND": "Expand",
